Ryddet i kode. La til dokumentasjon.
