+++ token-support-idporten/src/main/kotlin/no/nav/tms/token/support/idporten/authentication/AuthConfiguration.kt
@@ -1,0 +1,11 @@
+package no.nav.tms.token.support.idporten.authentication
+
+import com.auth0.jwk.JwkProvider
+
+internal class AuthConfiguration(
+        val contextPath: String,
+        val tokenCookieName: String,
+        val jwkProvider: JwkProvider,
+        val clientId: String,
+        val issuer: String
+)
