+++ token-support-idporten/src/main/kotlin/no/nav/tms/token/support/idporten/authentication/IdTokenPrincipal.kt
@@ -1,0 +1,6 @@
+package no.nav.tms.token.support.idporten.authentication
+
+import com.auth0.jwt.interfaces.DecodedJWT
+import io.ktor.auth.*
+
+internal data class IdTokenPrincipal(val decodedJWT: DecodedJWT) : Principal
