+++ token-support-idporten/src/main/kotlin/no/nav/tms/token/support/idporten/authentication/OauthServerConfigurationMetadata.kt
@@ -1,8 +1,9 @@
-package no.nav.tms.token.support.idporten
+package no.nav.tms.token.support.idporten.authentication
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
 
+// This object holds info returned from idporten's well-known-url
 @JsonIgnoreProperties(ignoreUnknown = true)
 internal data class OauthServerConfigurationMetadata(
         @JsonProperty(value = "issuer", required = true) val issuer: String,
