+++ token-support-idporten/src/main/kotlin/no/nav/tms/token/support/idporten/authentication/config/Environment.kt
@@ -1,4 +1,4 @@
-package no.nav.tms.token.support.idporten
+package no.nav.tms.token.support.idporten.authentication.config
 
 internal class Environment (
         val idportenWellKnownUrl: String = getIdportenEnvVar("IDPORTEN_WELL_KNOWN_URL"),
@@ -10,9 +10,4 @@ internal class Environment (
 private fun getIdportenEnvVar(varName: String): String {
     return System.getenv(varName)
             ?: throw IllegalArgumentException("Fant ikke $varName som brukes i token-support-idporten. Påse at nais.yaml er konfigurert riktig.")
-}
-
-private fun getEnvVar(varName: String, default: String = ""): String {
-    return System.getenv(varName)
-            ?: default
 }
