+++ token-support-idporten/src/main/kotlin/no/nav/tms/token/support/idporten/authentication/config/constants.kt
@@ -1,0 +1,8 @@
+package no.nav.tms.token.support.idporten.authentication.config
+
+internal object Idporten {
+    const val scope = "openid"
+    const val responseToken = "id_token"
+    const val authenticatorName = "tms_token_support_idporten"
+    const val postLoginRedirectCookie = "redirect_uri"
+}
