+++ token-support-idporten/src/main/kotlin/no/nav/tms/token/support/idporten/authentication/loginApi.kt
@@ -1,22 +1,27 @@
-package no.nav.tms.token.support.idporten
+package no.nav.tms.token.support.idporten.authentication
 
-import com.auth0.jwk.Jwk
 import com.auth0.jwt.JWT
-import com.auth0.jwt.algorithms.Algorithm
 import com.auth0.jwt.interfaces.DecodedJWT
-import com.auth0.jwt.interfaces.JWTVerifier
 import io.ktor.application.*
 import io.ktor.auth.*
 import io.ktor.http.*
 import io.ktor.response.*
 import io.ktor.routing.*
-import java.security.interfaces.RSAPublicKey
+import no.nav.tms.token.support.idporten.authentication.config.Idporten
+import no.nav.tms.token.support.idporten.authentication.config.RuntimeContext
 
 internal fun Routing.loginApi(runtimeContext: RuntimeContext) {
 
     val settings = runtimeContext.oauth2ServerSettings
 
     authenticate(Idporten.authenticatorName) {
+        // This method is empty because the authenticator will redirect any calls to idporten, which will in turn
+        // redirect to 'oath2/callback'. This method exists to differentiate between internal and external redirects
+        get("/oauth2/login") {}
+
+        // Users should arrive at this endpoint after a redirect from idporten, which will include a 'code' parameter
+        // This parameter will be used to retrieve the user's token directly from idporten, and will then be provided
+        // to the user as a token. The name of this token is determined when installing authentication.
         get("/oauth2/callback") {
             val principal = checkNotNull(call.authentication.principal<OAuthAccessTokenResponse.OAuth2>())
             when (val decodedJWT = settings.verify(principal, runtimeContext)) {
@@ -28,12 +33,10 @@ internal fun Routing.loginApi(runtimeContext: RuntimeContext) {
                 }
             }
         }
-
-        get("/oauth2/login") {}
     }
 }
 
-internal fun OAuthServerSettings.OAuth2ServerSettings.verify(tokenResponse: OAuthAccessTokenResponse.OAuth2?, runtimeContext: RuntimeContext): DecodedJWT? =
+private fun OAuthServerSettings.OAuth2ServerSettings.verify(tokenResponse: OAuthAccessTokenResponse.OAuth2?, runtimeContext: RuntimeContext): DecodedJWT? =
 tokenResponse?.idToken(Idporten.responseToken)?.let {
     runtimeContext.jwkProvider[JWT.decode(it).keyId].idTokenVerifier(
             clientId,
@@ -41,12 +44,4 @@ tokenResponse?.idToken(Idporten.responseToken)?.let {
     ).verify(it)
 }
 
-internal fun OAuthAccessTokenResponse.OAuth2.idToken(tokenCookieName: String): String? = extraParameters[tokenCookieName]
-
-internal fun Jwk.idTokenVerifier(clientId: String, issuer: String): JWTVerifier =
-        JWT.require(this.RSA256())
-                .withAudience(clientId)
-                .withIssuer(issuer)
-                .build()
-
-internal fun Jwk.RSA256() = Algorithm.RSA256(publicKey as RSAPublicKey, null)
+private fun OAuthAccessTokenResponse.OAuth2.idToken(tokenCookieName: String): String? = extraParameters[tokenCookieName]
