--- token-support-idporten/src/main/kotlin/no/nav/tms/token/support/idporten/constants.kt
@@ -1,8 +1,0 @@
-package no.nav.tms.token.support.idporten
-
-internal object Idporten {
-    val scope = "openid"
-    val responseToken = "id_token"
-    val authenticatorName = "tms_token_support_idporten"
-    val postLoginRedirectCookie = "redirect_uri"
-}
