+++ token-support-idporten/src/main/kotlin/no/nav/tms/token/support/idporten/user/IdportenUser.kt
@@ -1,18 +1,22 @@
 package no.nav.tms.token.support.idporten.user
 
+import com.auth0.jwt.interfaces.Claim
+import com.auth0.jwt.interfaces.DecodedJWT
 import java.time.Instant
 import java.time.temporal.ChronoUnit
 
 data class IdportenUser (
-    val ident: String,
-    val loginLevel: Int,
-    val token: String,
-    val tokenExpirationTime: Instant,
-    val auxiliaryEssoToken: String? = null
+        val ident: String,
+        val loginLevel: Int,
+        val tokenExpirationTime: Instant,
+        val jwt: DecodedJWT
 ) {
+    val tokenString get() = jwt.token
+
+    val tokenClaims get() = jwt.claims
 
     fun createAuthenticationHeader(): String {
-        return "Bearer $token"
+        return "Bearer $tokenString"
     }
 
     override fun toString(): String {
